requirements.txt
fastapi
httpx
pydantic
uvicorn
python-dotenv
click
